Task5 calculators

See merge request AXEL.ROJASRETANA/paralela25a-stockholm_syndrome!13

+++ web_server_11/src/webapp/calculator/GoldbachCalculator.cpp
@@ -2,7 +2,7 @@
 
 #include "GoldbachCalculator.hpp"
 
-int64_t GoldbachCalculator::processNumber(int64_t number,
+size_t GoldbachCalculator::processNumber(int64_t number,
     std::vector<int64_t>& goldbachSums) {
     goldbachSums.push_back(number);
     int64_t sumOperands = 1;
